+++ src/api/index.ts
@@ -170,26 +170,32 @@ export const start = async (): Promise<void> => {
         )}`;
       } catch (error) {
         if (error instanceof RateLimiterRes) {
-          logger.warn(
-            "rate-limiter",
-            `${rateLimitKey} ${apiKey?.appName} reached allowed rate limit ${
-              rateLimitRule.options.points
-            } requests in ${rateLimitRule.options.duration}s by calling ${
-              error.consumedPoints
-            } times in rule ${JSON.stringify(rateLimitRule)}`
-          );
+          if (
+            error.consumedPoints == Number(rateLimitRule.options.points) + 1 ||
+            error.consumedPoints % 50 == 0
+          ) {
+            logger.warn(
+              "rate-limiter",
+              `${rateLimitKey} ${apiKey?.appName} reached allowed rate limit ${
+                rateLimitRule.options.points
+              } requests in ${rateLimitRule.options.duration}s by calling ${
+                error.consumedPoints
+              } times in rule ${JSON.stringify(rateLimitRule)}`
+            );
+          }
 
-          const tooManyRequestsResponse = {
-            statusCode: 429,
-            error: "Too Many Requests",
-            message: `Max ${rateLimitRule.options.points} requests in ${rateLimitRule.options.duration}s reached`,
-          };
-
-          return reply
-            .response(tooManyRequestsResponse)
-            .type("application/json")
-            .code(429)
-            .takeover();
+          // const tooManyRequestsResponse = {
+          //   statusCode: 429,
+          //   error: "Too Many Requests",
+          //   message: `Max ${rateLimitRule.options.points} requests in ${rateLimitRule.options.duration}s reached`,
+          // };
+          //
+          // return reply
+          //   .response(tooManyRequestsResponse)
+          //   .header("x-cf-block", "true")
+          //   .type("application/json")
+          //   .code(429)
+          //   .takeover();
         } else {
           throw error;
         }